+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.0.2 (unreleased)
+------------------
+- Removed default parameter
+- Added replacing parameters in translation strings
+
 0.0.1
 -----
 - Added basic functions